Support pre-fork via feersum/plackup scripts

+++ lib/Feersum/Runner.pm
@@ -5,6 +5,7 @@ use EV;
 use Feersum;
 use Socket qw/SOMAXCONN/;
 use POSIX ();
+use Scalar::Util qw/weaken/;
 
 sub new { my $c = shift; bless {@_},$c }
 
@@ -34,28 +35,32 @@ sub _prepare {
     my $f = Feersum->endjinn;
     $f->use_socket($sock);
 
+    if ($self->{options}) {
+        # Plack::Runner puts these here
+        $self->{pre_fork} = delete $self->{options}{pre_fork};
+        $self->{quiet} = 1;
+    }
+
     $self->{endjinn} = $f;
+}
+
+# for overriding:
+sub assign_request_handler {
+    $_[0]->{endjinn}->request_handler($_[1]);
 }
 
 sub run {
     my $self = shift;
+    weaken $self;
 
     $self->{quiet} or warn "Feersum [$$]: starting...\n";
     $self->_prepare();
 
     my $rh = shift || delete $self->{app} || do $self->{app_file};
-    die "not a code ref" unless ref($rh) eq 'CODE';
-    $self->{endjinn}->request_handler($rh);
+    $self->assign_request_handler($rh);
     undef $rh;
 
-    $self->{_quit} = EV::signal 'QUIT', sub {
-        $self->{_shutdown} = 1;
-        $self->{quiet} or warn "Feersum [$$]: got SIGQUIT!\n";
-        undef $self->{_quit};
-        $SIG{QUIT} = 'IGNORE';
-        $self->{endjinn}->graceful_shutdown(sub { POSIX::exit(0) });
-        $self->{_death} = EV::timer 5, 0, sub { POSIX::exit(1) };
-    };
+    $self->{_quit} = EV::signal 'QUIT', sub { $self->quit };
 
     $self->pre_fork if $self->{pre_fork};
     EV::loop;
@@ -64,6 +69,7 @@ sub run {
 
 sub fork_another {
     my ($self, $slot) = @_;
+    weaken $self;
 
     my $pid = fork;
     die "failed to fork: $!" unless defined $pid;
@@ -101,15 +107,27 @@ sub pre_fork {
     $self->fork_another($_) for (1 .. $self->{pre_fork});
 
     $self->{endjinn}->unlisten();
-    # broadcast SIGQUIT to the group
-    $self->{_quit} = EV::signal 'QUIT', sub {
-        $self->{_shutdown} = 1;
-        $self->{quiet} or warn "Feersum [$$]: got SIGQUIT!\n";
-        undef $self->{_quit};
-        $SIG{QUIT} = 'IGNORE';
+}
+
+sub quit {
+    my $self = shift;
+    return if $self->{_shutdown};
+
+    $self->{_shutdown} = 1;
+    $self->{quiet} or warn "Feersum [$$]: shutting down...\n";
+    my $death = 5;
+
+    if ($self->{_n_kids}) {
+        # in parent, broadcast SIGQUIT to the group
         kill 3, -$$; # kill process group, but not self
-        $self->{_death} = EV::timer 7, 0, sub { POSIX::exit(1) };
-    };
+        $death += 2;
+    }
+    else {
+        # in child or solo process
+        $self->{endjinn}->graceful_shutdown(sub { POSIX::exit(0) });
+    }
+
+    $self->{_death} = EV::timer $death, 0, sub { POSIX::exit(1) };
 }
 
 1;
@@ -123,14 +141,43 @@ Feersum::Runner
 
     use Feersum::Runner;
     my $runner = Feersum::Runner->new(
-        listen => 'localhost:12345',
+        listen => 'localhost:5000',
+        pre_fork => 0,
+        quiet => 1,
+        app_file => 'app.feersum',
     );
     $runner->run($feersum_app);
 
 =head1 DESCRIPTION
 
-Much like L<Plack::Runner>, but with far fewer options (only a single 'listen'
-or 'host'/'port' pair is currently supported).
+Much like L<Plack::Runner>, but with far fewer options.
+
+=head1 ATTRIBUTES
+
+The following initialization options are available.
+
+=over 4
+
+=item listen
+
+Listen on this TCP socket (C<host:port> format).
+
+=item pre_fork
+
+Fork this many worker processes.
+
+The fork is run immediately at startup and after the app is loaded (i.e. in
+the C<run()> method).
+
+=item quiet
+
+Don't be so noisy.
+
+=item app_file
+
+Load this filename as a native feersum app.
+
+=back
 
 =head1 METHODS
